requirements.txt
streamlit==1.44.1
autogen==0.8.5
pinecone==6.0.2
google-generativeai==0.8.4
sentence-transformers==2.2.2
transformers==4.30.2
huggingface-hub==0.16.4


